move consul/vault template functions into library
Previously Consul and Vault template functions were always included in
the template's function maps. This removes them as defaults and adds
them to the library of all available template functions.

Mostly just migration with the necessary tweaks.
