+++ tfunc/deny_test.go
@@ -1,0 +1,13 @@
+package tfunc
+
+import "testing"
+
+func TestDeny(t *testing.T) {
+	v, err := DenyFunc()
+	if v != "" {
+		t.Errorf("bad return string: '%v'", v)
+	}
+	if err != disabledErr {
+		t.Errorf("bad error: %v", err)
+	}
+}
